Refactor pop-up windows and add account options
This commit refactors the pop-up windows in the index.html file and adds an account options pop-up. The pop-up windows now have a consistent class name "popup_window" and use the "grow" animation. The new account options pop-up is positioned in the center of the screen and displays options for signing out and deleting the account.

Code changes:
- Updated the index.html file to add the account options pop-up and update the class names of the existing pop-up windows.
- Updated the global.css file to add styles for the account options pop-up and remove unused styles.